+++ src/app/(afterLogin)/layout.tsx
@@ -1,34 +1,27 @@
-"use client";
-
 import Link from "next/link";
 import React from "react";
-import { usePathname, useSelectedLayoutSegment } from "next/navigation";
-
-import { styled } from "styled-components";
+import { auth } from "@/auth";
 
 import Nav from "./_component/nav";
-import FixedSearchBar from "./_component/SearchBar";
-import TrendsForYou from "./_component/TrendsForYou";
-import WhoToFollow from "./_component/WhoToFollow";
 import NavProfile from "./_component/nav-profile";
-import SearchFilter from "./search/_component/SearchFilter";
+import LayoutRightSection from "./_component/right-seciton";
+import { Container, Header, FixedHeader, MainLogo } from "./layout-css";
 
-const AfterLoginLayout = ({
+const AfterLoginLayout = async ({
   children,
   modal,
 }: Readonly<{
   children: React.ReactNode;
   modal: React.ReactNode;
 }>) => {
-  const segment = useSelectedLayoutSegment();
-
-  const pathname = usePathname();
-
+  const session = await auth();
+  console.log("ajskndkjand", session);
+  const path = {};
   return (
     <Container>
       <Header>
         <FixedHeader>
-          <Link href="/home">
+          <Link href={session?.user ? "/home" : "/"}>
             <MainLogo>
               <svg viewBox="0 0 24 24" aria-hidden="true" width="20rem" height="2.5rem">
                 <g>
@@ -37,193 +30,14 @@ const AfterLoginLayout = ({
               </svg>
             </MainLogo>
           </Link>
-          <Nav></Nav>
-          <NavProfile />
+          <Nav me={session} />
+          <NavProfile me={session} />
         </FixedHeader>
       </Header>
-      {segment === "messages" ? (
-        <MessageMain>{children}</MessageMain>
-      ) : (
-        <RightSection>
-          <RightSectionInner>
-            <Main>{children}</Main>
-            <Right>
-              {pathname === "/search" || pathname === "/explore" ? null : (
-                <SearchConatiner>
-                  <FixedSearchBar></FixedSearchBar>
-                </SearchConatiner>
-              )}
-              <SearchFilter></SearchFilter>
-              <TrendsForYou></TrendsForYou>
-              <WhoToFollow></WhoToFollow>
-            </Right>
-          </RightSectionInner>
-        </RightSection>
-      )}
+      <LayoutRightSection me={session}>{children}</LayoutRightSection>
       {modal}
     </Container>
   );
 };
 
-const Container = styled.div`
-  display: flex;
-  align-items: stretch;
-  justify-content: center;
-  background-color: #fff;
-`;
-
-const Header = styled.header`
-  box-sizing: border-box;
-  width: 275px;
-  /* flex-grow: 1; */
-  display: flex;
-  justify-content: flex-end;
-
-  @media (min-width: 1295px) {
-    justify-content: center;
-  }
-`;
-
-const FixedHeader = styled.div`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 88px;
-  position: fixed;
-  height: 100%;
-  overflow-x: hidden;
-  padding: 0 8px;
-  background-color: white;
-
-  @media (min-width: 1295px) {
-    width: 275px;
-    align-items: flex-start;
-  }
-`;
-
-const MainLogo = styled.div`
-  width: 52px;
-  height: 52px;
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  &:hover {
-    background-color: rgb(15, 20, 25, 0.1);
-    border-radius: 50%;
-  }
-`;
-
-const Profile = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 12px;
-  box-sizing: border-box;
-  width: inherit;
-  cursor: pointer;
-
-  &:hover {
-    background-color: rgb(15, 20, 25, 0.1);
-    border-radius: 9999px;
-  }
-  & > div {
-    height: 40px;
-  }
-
-  & > div:nth-child(1) {
-    width: 40px;
-    border-radius: 50%;
-    background-color: red;
-  }
-
-  & > div:nth-child(2) {
-    display: none;
-  }
-
-  svg {
-    display: none;
-  }
-
-  @media (min-width: 1295px) {
-    & > div:nth-child(2) {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-      justify-content: center;
-      max-width: 100%;
-      margin: 12px;
-      flex-grow: 1;
-
-      div:nth-child(2) {
-        color: rgb(83, 100, 113);
-      }
-    }
-    svg {
-      display: inline-block;
-    }
-  }
-`;
-
-const RightSection = styled.main`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  width: 600px;
-  height: 100dvh;
-
-  @media (min-width: 1024px) {
-    width: 1050px;
-  }
-`;
-
-const RightSectionInner = styled.div`
-  height: 100%;
-  width: 600px;
-
-  display: flex;
-  justify-content: space-between;
-
-  @media (min-width: 1024px) {
-    width: 990px;
-  }
-`;
-
-const Right = styled.div`
-  display: none;
-
-  @media (min-width: 1024px) {
-    display: flex;
-    width: 350px;
-    height: 100dvh;
-    box-sizing: border-box;
-    margin-right: 10px;
-    /* position: relative; */
-
-    flex-direction: column;
-  }
-`;
-
-const SearchConatiner = styled.div`
-  display: flex;
-
-  justify-content: start;
-  width: 350px;
-  height: 53px;
-  box-sizing: border-box;
-  margin-bottom: 20px;
-`;
-
-const Main = styled.div`
-  box-sizing: border-box;
-  width: 600px;
-  height: 200dvh;
-`;
-
-const MessageMain = styled.div`
-  box-sizing: border-box;
-  width: 1050px;
-`;
 export default AfterLoginLayout;
