+++ src/app/(beforeLogin)/_component/SignupStyled.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import styled from "styled-components";
+
+const ModalBody = styled.div`
+  box-sizing: border-box;
+  width: 100%;
+  /* height: 100%; */
+
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  /* justify-content: space-between; */
+  /* align-items:space-between */
+
+  padding: 0 80px;
+`;
