+++ src/app/(beforeLogin)/page.tsx
@@ -1,7 +1,10 @@
-"use client";
-
 import Home from "@/app/(beforeLogin)/_component/Home";
 
 export default function HomePage() {
-  return <Home />;
+  return (
+    <>
+      {/* @ts-expect-error Async Server Component */}
+      <Home />
+    </>
+  );
 }
