+++ src/mocks/handlers.ts
@@ -11,6 +11,7 @@ function generateDate() {
 }
 const User = [
   { id: "elonmusk", nickname: "Elon Musk", image: "/yRsRRjGO.jpg" },
+  { id: "surrrrfing", nickname: "김지환", image: faker.image.avatar() },
   { id: "zerohch0", nickname: "제로초", image: "/5Udwvqim.jpg" },
   { id: "leoturtle", nickname: "레오", image: faker.image.avatar() },
 ];
